+++ test/config.test.js
@@ -798,7 +798,7 @@ describe(`configTest: ${printPath("[test/config.test.js]")}`, function () {
                     connectionURI: "http://localhost:8080",
                 },
                 appInfo: {
-                    apiDomain: "http://api.supertokens.io",
+                    apiDomain: "https://api.supertokens.io",
                     appName: "SuperTokens",
                     websiteDomain: "supertokens.io",
                     apiBasePath: "test/",
@@ -809,7 +809,7 @@ describe(`configTest: ${printPath("[test/config.test.js]")}`, function () {
 
             assert(SessionRecipe.getInstanceOrThrowError().config.antiCsrf === "NONE");
             assert(SessionRecipe.getInstanceOrThrowError().config.cookieSameSite === "lax");
-            assert(SessionRecipe.getInstanceOrThrowError().config.cookieSecure === false);
+            assert(SessionRecipe.getInstanceOrThrowError().config.cookieSecure === true);
             resetAll();
         }
 
@@ -1066,6 +1066,25 @@ describe(`configTest: ${printPath("[test/config.test.js]")}`, function () {
                     throw err;
                 }
             }
+            resetAll();
+        }
+
+        {
+            STExpress.init({
+                supertokens: {
+                    connectionURI: "http://localhost:8080",
+                },
+                appInfo: {
+                    apiDomain: "https://localhost",
+                    appName: "Supertokens",
+                    websiteDomain: "http://localhost:3000",
+                },
+                recipeList: [Session.init()],
+            });
+
+            assert(SessionRecipe.getInstanceOrThrowError().config.cookieSecure);
+            assert(SessionRecipe.getInstanceOrThrowError().config.cookieSameSite === "none");
+
             resetAll();
         }
     });