+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Changes implementation such that actual client IDs are not in the SDK, removes imports for OAuth dev related code.
 
+### Fixed
+
+-   URL protocol is being taken into account when determining the value of cookie same site: https://github.com/supertokens/supertokens-golang/issues/36
+
 ## [7.3.0] - 2021-10-11
 
 ### Added